:zap hahaha

+++ micro-services/cliente-service/src/main/java/br/com/tads/dac/clienteservice/model/dto/MilesTranscationResponseDTO.java
@@ -7,5 +7,5 @@ import lombok.Data;
 @Builder
 public class MilesTranscationResponseDTO {
     private String codigo;
-    private int saldoMilhas;
+    private Long saldoMilhas;
 }
